requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
cffi==1.17.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.0.0
Django==4.2.17
django-bootstrap-v5==1.0.11
django-extensions==3.2.3
django-health-check==3.18.3
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
packaging==24.2
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
python-decouple==3.8
python-dotenv==1.0.1
service-identity==24.2.0
setuptools==75.8.0
soupsieve==2.6
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
zope.interface==7.2
